Updates to xmcd_visualiser
## xmcd_visualiser/__init__.py
 - refactor tkgui code into a seperate pacakge

## average_plot.py
 - add plots of all selected XAS spectra
