+++ mmg_toolbox/tkguis/xmcd_visualiser/__init__.py
@@ -1,0 +1,5 @@
+
+from .widget import XMCDVisualiser
+from .app import create_xmcd_visualiser
+
+__all__ = ['XMCDVisualiser', 'create_xmcd_visualiser']
